ex2_code_na_lab: IO weggehaald van check uit JIWY.

+++ Ex2/ex2_code_na_lab/JIWYrealtime/JIWY/include/JIWY/JIWY.h
@@ -29,7 +29,7 @@ class JIWY : public Parallel
 
 public:
   // Define constructor and destructor
-  JIWY(ChannelOut<double> *joystick_hor_in, ChannelOut<double> *joystick_vert_in, ChannelOut<double> *robot_hor_feedback, ChannelIn<uint16_t> *robot_hor_out, ChannelOut<double> *robot_ver_feedback, ChannelIn<uint16_t> *robot_ver_out);
+  JIWY(ChannelOut<double> *joystick_hor_in, ChannelOut<double> *joystick_vert_in, ChannelOut<double> *robot_hor_feedback, ChannelIn<double> *robot_hor_out, ChannelOut<double> *robot_ver_feedback, ChannelIn<double> *robot_ver_out);
   virtual ~JIWY();
 
 private:
